requirements.txt
fastapi
uvicorn
openai-whisper
requests
torch
